In-repo typings file structure
