+++ src/test/typescript/io-spec.ts
@@ -1,4 +1,4 @@
-import { IO } from 'monet';
+import { IO } from '../../../index';
 
 const logX = IO(() => console.log('X'));
 const getY = logX.map(() => 'Y');
